+++ player2/player/frame_parser/frame_parser_video_vc1.cpp
@@ -523,8 +523,8 @@ FrameParserStatus_t   FrameParser_VideoVc1_c::RevPlayGeneratePostDecodeParameter
 
         if (!ParsedFrameParameters->ReferenceFrame)
         {
-            ReverseQueuedPostDecodeSettingsRing->Insert ((unsigned int)ParsedFrameParameters);
-            ReverseQueuedPostDecodeSettingsRing->Insert ((unsigned int)ParsedVideoParameters);
+            ReverseQueuedPostDecodeSettingsRing->Insert ((uintptr_t)ParsedFrameParameters);
+            ReverseQueuedPostDecodeSettingsRing->Insert ((uintptr_t)ParsedVideoParameters);
         }
         else
 
@@ -537,8 +537,8 @@ FrameParserStatus_t   FrameParser_VideoVc1_c::RevPlayGeneratePostDecodeParameter
 
             while (ReverseQueuedPostDecodeSettingsRing->NonEmpty())
             {
-                ReverseQueuedPostDecodeSettingsRing->Extract ((unsigned int *)&DeferredParsedFrameParameters);
-                ReverseQueuedPostDecodeSettingsRing->Extract ((unsigned int *)&DeferredParsedVideoParameters);
+                ReverseQueuedPostDecodeSettingsRing->Extract ((uintptr_t *)&DeferredParsedFrameParameters);
+                ReverseQueuedPostDecodeSettingsRing->Extract ((uintptr_t *)&DeferredParsedVideoParameters);
                 CalculateFrameIndexAndPts (DeferredParsedFrameParameters, DeferredParsedVideoParameters);
             }
         }
